+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.2]
+
+## Changed
+- Fire `TruexAdFree` in the `adBreakFinished` listener instead of `adFinished`, as the stream isn't fully reloaded for seeking after `adFinished`
+- Update Bitmovin Web SDK to 08.31.0
+- `player.isLive()` returns false when in a VPAID, so store `isLiveStream` in a variable upon playing the stream.
+
 ## [1.2.0]
 
 ## Added
@@ -25,6 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial yospace integration
 
+[1.2.2]: https://github.com/bitmovin/bitmovin-player-web-integrations-yospace/compare/1.2.1...1.2.2
 [1.2.0]: https://github.com/bitmovin/bitmovin-player-web-integrations-yospace/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/bitmovin/bitmovin-player-web-integrations-yospace/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/bitmovin/bitmovin-player-web-integrations-yospace/compare/1406694e8e7c63bfa9d24b84fd6253b135cc0e74...1.0.0